% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add_unimodUI}
\alias{add_unimodUI}
\title{User interface of MGF parameters}
\usage{
add_unimodUI(
  id,
  sites = c("A", "C", "D", "E", "F", "G", "H", "I", "K", "L", "M", "N", "P", "Q", "R",
    "S", "T", "W", "Y", "V", "U", "B", "X", "Z", "N-term", "C-term"),
  positions = c("Anywhere", "Any N-term", "Protein N-term", "Any C-term",
    "Protein C-term")
)
}
\arguments{
\item{id}{Namespace identifier.}

\item{sites}{Sites of amino-acid residues.}

\item{positions}{Positions of modifications.}
}
\description{
User interface of MGF parameters
}
