% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{fastaServer}
\alias{fastaServer}
\title{Server-side processing of FASTA parameters}
\usage{
fastaServer(
  id,
  accessions = c("uniprot_acc", "uniprot_id", "refseq_acc", "other")
)
}
\arguments{
\item{id}{Namespace identifier.}

\item{accessions}{Accession types.}
}
\description{
Server-side processing of FASTA parameters
}
