% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{fastaUI0}
\alias{fastaUI0}
\title{User interface of FASTA parameters}
\usage{
fastaUI0(
  id,
  accessions = c("uniprot_acc", "uniprot_id", "refseq_acc", "other")
)
}
\arguments{
\item{id}{Namespace identifier.}

\item{accessions}{Protein accession types.}
}
\description{
User interface of FASTA parameters
}
