% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2map.R
\name{map_ms2Server}
\alias{map_ms2Server}
\title{Server-side processing of MGF parameters}
\usage{
map_ms2Server(id, type_ms2ions = c("by", "ax", "cz"))
}
\arguments{
\item{id}{Namespace identifier.}

\item{type_ms2ions}{Type of MS2 ions.}
}
\description{
Server-side processing of MGF parameters
}
