% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2map.R
\name{map_ms2UI}
\alias{map_ms2UI}
\title{User interface of MGF parameters}
\usage{
map_ms2UI(id, type_ms2ions = c("by", "ax", "cz"))
}
\arguments{
\item{id}{Namespace identifier.}

\item{type_ms2ions}{Type of MS2 ions.}
}
\description{
User interface of MGF parameters
}
