% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modification.R
\name{modServer}
\alias{modServer}
\title{Server-side processing of search parameters}
\usage{
modServer(id)
}
\arguments{
\item{id}{Namespace identifier.}
}
\description{
Server-side processing of search parameters
}
