% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchServer}
\alias{searchServer}
\title{Server-side processing of search parameters}
\usage{
searchServer(id)
}
\arguments{
\item{id}{Namespace identifier.}
}
\description{
Server-side processing of search parameters
}
