% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{find_unimodServer}
\alias{find_unimodServer}
\title{Server-side processing of MGF parameters}
\usage{
find_unimodServer(id)
}
\arguments{
\item{id}{Namespace identifier.}
}
\description{
Server-side processing of MGF parameters
}
