% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{remove_unimodServer}
\alias{remove_unimodServer}
\title{Server-side processing of MGF parameters}
\usage{
remove_unimodServer(id, temp_dir = NULL, file_umods = "unimods.txt")
}
\arguments{
\item{id}{Namespace identifier.}
}
\description{
Server-side processing of MGF parameters
}
